Moved task classes under StateChange classes that generate them
