+++ engine/src/main/java/org/terasology/rendering/dag/stateChanges/BindFBO.java
@@ -22,12 +22,41 @@ import org.terasology.rendering.opengl.FBOManagerSubscriber;
 import com.google.common.base.Objects;
 import org.terasology.rendering.dag.RenderPipelineTask;
 import org.terasology.rendering.dag.StateChange;
-import org.terasology.rendering.dag.tasks.BindFBOTask;
+
+import static org.lwjgl.opengl.EXTFramebufferObject.GL_FRAMEBUFFER_EXT;
+import static org.lwjgl.opengl.EXTFramebufferObject.glBindFramebufferEXT;
 
 /**
  * TODO: Add javadocs
  */
 public final class BindFBO implements FBOManagerSubscriber, StateChange {
+
+    private final class BindFBOTask implements RenderPipelineTask {
+
+        private int fboId;
+        private final ResourceUrn fboName;
+
+        private BindFBOTask(int fboId, ResourceUrn fboName) {
+            this.fboId = fboId;
+            this.fboName = fboName;
+        }
+
+        @Override
+        public void execute() {
+            glBindFramebufferEXT(GL_FRAMEBUFFER_EXT, fboId);
+        }
+
+        private void setFboId(int fboId) {
+            this.fboId = fboId;
+        }
+
+        @Override
+        public String toString() {
+            return String.format("%30s: %s (fboId:%s)", this.getClass().getSimpleName(), fboName, fboId);
+        }
+    }
+
+
     private static final Integer DEFAULT_FRAME_BUFFER_ID = 0;
     // TODO: add necessary checks for ensuring generating FBO with the name "display" is not possible.
     private static final ResourceUrn DEFAULT_FRAME_BUFFER_URN = new ResourceUrn("engine:display");
