+++ engine/src/main/java/org/terasology/rendering/dag/stateChanges/LookThrough.java
@@ -20,7 +20,6 @@ import org.terasology.rendering.cameras.Camera;
 import org.terasology.rendering.dag.RenderPipelineTask;
 import org.terasology.rendering.dag.StateChange;
 import org.terasology.rendering.dag.tasks.LookThroughDefaultCameraTask;
-import org.terasology.rendering.dag.tasks.LookThroughTask;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
@@ -35,6 +34,31 @@ import static com.google.common.base.Preconditions.checkNotNull;
  * The default instance of this class resets both matrices to identity matrices, opengl's default.
  */
 public class LookThrough implements StateChange {
+
+    private class LookThroughTask implements RenderPipelineTask {
+
+        private Camera camera;
+
+        /**
+         * Constructs an instance of this class initialized with the given camera.
+         *
+         * @param camera an instance implementing the Camera interface
+         */
+        private LookThroughTask(Camera camera) {
+            this.camera = camera;
+        }
+
+        @Override
+        public void execute() {
+            camera.lookThrough();
+        }
+
+        @Override
+        public String toString() {
+            return String.format("%30s: %s", this.getClass().getSimpleName(), camera.toString());
+        }
+    }
+
 
     private static LookThrough defaultInstance = new LookThrough();
