+++ engine/src/main/java/org/terasology/rendering/dag/stateChanges/SetInputTexture.java
@@ -16,11 +16,17 @@
 package org.terasology.rendering.dag.stateChanges;
 
 import org.terasology.assets.ResourceUrn;
+import org.terasology.rendering.assets.material.Material;
 import org.terasology.rendering.dag.RenderPipelineTask;
 import org.terasology.rendering.dag.StateChange;
-import org.terasology.rendering.dag.tasks.SetInputTextureTask;
 
 import java.util.Objects;
+
+import static org.lwjgl.opengl.GL11.GL_TEXTURE_2D;
+import static org.lwjgl.opengl.GL11.glBindTexture;
+import static org.lwjgl.opengl.GL13.GL_TEXTURE0;
+import static org.lwjgl.opengl.GL13.glActiveTexture;
+import static org.terasology.rendering.dag.AbstractNode.getMaterial;
 
 /**
  * This StateChange generates the tasks that set and reset input textures.
@@ -29,6 +35,46 @@ import java.util.Objects;
  * This StateChange and the underlying task only handles textures of type GL_TEXTURE_2D.
  */
 public class SetInputTexture implements StateChange {
+
+    /**
+     * Instances of this class bind a texture to a texture unit. The integer identifying
+     * the texture unit is then passed to a shader program using the material/parameter
+     * pair provided on construction. See the source of the execute() method for the
+     * nitty gritty details.
+     *
+     * WARNING: RenderPipelineTasks are not meant for direct instantiation and manipulation.
+     * Modules or other parts of the engine should take advantage of them through classes
+     * inheriting from StateChange.
+     */
+    private class SetInputTextureTask implements RenderPipelineTask {
+
+        private final int textureSlot;
+        private final int textureId;
+        private final Material material;
+        private final String materialParameter;
+
+        private SetInputTextureTask(int textureSlot, int textureId, ResourceUrn materialURN, String materialParameter) {
+            this.textureSlot = textureSlot;
+            this.textureId = textureId;
+            this.material = getMaterial(materialURN);
+            this.materialParameter = materialParameter;
+        }
+
+        @Override
+        public void execute() {
+            glActiveTexture(GL_TEXTURE0 + textureSlot);
+            glBindTexture(GL_TEXTURE_2D, textureId);
+            material.setInt(materialParameter, textureSlot, true);
+        }
+
+        @Override
+        public String toString() {
+            return String.format("%30s: slot %s, texture %s, material %s, parameter %s", this.getClass().getSimpleName(),
+                    textureSlot, textureId, material.getUrn().toString(), materialParameter);
+        }
+
+    }
+
 
     private final int textureSlot;
     private final int textureId;
