--- engine/src/main/java/org/terasology/rendering/dag/tasks/EnableMaterialTask.java
@@ -1,42 +1,0 @@
-/*
- * Copyright 2016 MovingBlocks
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.terasology.rendering.dag.tasks;
-
-import org.terasology.rendering.assets.material.Material;
-import org.terasology.rendering.dag.RenderPipelineTask;
-
-/**
- * TODO: Add javadocs
- */
-public final class EnableMaterialTask implements RenderPipelineTask {
-    private Material material;
-    private String materialName;
-
-    public EnableMaterialTask(Material material, String materialName) {
-        this.material = material;
-        this.materialName = materialName;
-    }
-
-    @Override
-    public void execute() {
-        material.enable();
-    }
-
-    @Override
-    public String toString() {
-        return String.format("%30s: %s", this.getClass().getSimpleName(), materialName);
-    }
-}
